app/requirements.txt
# Flask Web Application Requirements
flask==3.0.0
flask-cors==4.0.0
flask-session==0.5.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Security
bcrypt==4.1.2
werkzeug==3.0.1

# PDF Export
reportlab==4.0.7

# QR Code
qrcode==7.4.2
Pillow==10.1.0
pyzbar==0.1.9
opencv-python==4.10.0.84

# Utilities
numpy==2.4.2
python-dateutil==2.8.2
